Fix Java Mail Store.connect hangs without timeout
Add an worker level timeout to avoid the worker hanging because of
a possible dropped connection on the server side that would not
generate a timeout error.

The symptom is a connection that stayed established forever with no
timeout even though timeouts have been defined for the connection and
the socket.

Issue similar to the one described here:
https://community.oracle.com/thread/2487900